chore:move-receive-payments-to-account settings
