+++ src/routes/_dashboard-layout/dashboard/account/receiving-payments.tsx
